The user can now choose if they want to border or background to be highlighted on key press
The user can change these settings in the config.txt. I also added a helper function for grabbing specific lines of the config.txt. The reason why is because the other way I did it was tedious and hard for me to wrap my head around, but this is pretty straightforward.